Use path interface for S3 buckets.
S3 used to work by providing a path to bucket names, and
then they started supporting a nice API via URLs.  While
that tends to work well overall, it has an important
disavantage which is not working with alternative
implementations which provide an IP address only.

This change should remain compatible with Amazon's S3,
while offering the chance of interacting with the API
through an IP address.