An S3 integration test relying on SigV4 fix.
This integration test issues a request that actually uses a query
argument.  It fails (with an authentication error) without the fix from
the previous revision and passes with it.

This serves as an integration test for the codepath through the
authentication code being fixed (so we don't rely only on unit tests).