test_ssl_hostname_verification honors its docstring.
It asserts that the created Agent received an instance of
twisted.internet.ssl.VerifyingContextFactory.  This fixes an issue
caused by Twisted 17.1.0 and later's use of HostnameEndpoint within
t.w.client.Agent.

This test should be deleted as part of the deprecation process
described in #47 (TLS support includes root certificate management
code that should be handled elsewhere).